feat: ListResponse.Resources discriminator
This allows pydantic to guess the models to use when parsing a
ListResponse payload. For the moment only User and Group and
ServiceProviderConfig are in the
union, because there are some bugs with the implementation EnterpriseUser and ResourceType,
and the Schema implementation is missing.

This needs pydantic 2.5 due to the usage of pydantic.Discriminator

This needs python 3.9 due to the usage of typing.Annotated